DjangoBackend/requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.8
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
filelock==3.15.4
firebase-admin==6.5.0
flake8==7.1.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==23.0.0
httplib2==0.22.0
identify==2.6.0
idna==3.7
mccabe==0.7.0
msgpack==1.0.8
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pre-commit==3.7.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3